Implement motion detection
changes also includes using MVVM on HomeViewController using HomeVIewModel
